+++ UI/Controls/ValueEditors/SpriteAnimationEditor.xaml.cs
@@ -22,7 +22,7 @@
 
         public SpriteAnimationEditor() {
             this.SelectSpriteAnimationCommand = new RelayCommand(() => {
-                if (this._dialogService.ShowSelectAssetDialog(this._projectService.CurrentProject, AssetType.SpriteAnimation, AssetType.SpriteAnimation, true, out var asset)) {
+                if (this._dialogService.ShowSelectAssetDialog(this._projectService.CurrentProject, typeof(SpriteAnimation), true, out var asset)) {
                     this.Asset = asset as SpriteAnimationAsset;
                     this.Value = this.Asset?.SavableValue;
                 }